+++ server/routers/authRouter.js
@@ -1,0 +1,72 @@
+import jwt from 'jsonwebtoken';
+import bcrypt from 'bcrypt';
+import express from 'express';
+import { Router } from 'express';
+import User from '../models/User.js';
+
+const app = express();
+app.use(express.json());
+const authRouter = Router();
+
+//AUTHENTICATION ROUTES
+authRouter.post('/signup', async (req, res) => {
+    try {
+        // Check if user already exists
+        const existingUser = await User.findOne({ email: req.body.email.toLowerCase(), });
+        if (existingUser) {
+            return res.status(400).json({ error: 'User already exists' });
+        }
+        // Encrypt password to store in the db
+        const passwordHash = bcrypt.hashSync(req.body.password, 10);
+        const newUser = new User({
+            ...req.body,
+            password: passwordHash,
+            role: null,
+            status: 'pending'
+        });
+        await newUser.save();
+        // Create a sessionToken that is unique to one person _id and JWT Secret
+        const sessionToken = jwt.sign(
+            { _id: newUser._id },
+            process.env.JWT_SECRET,
+            { expiresIn: 60 * 60 * 24 }
+        );
+        res.json({ message: 'Sign up successful', sessionToken });
+    } catch (err) {
+        res.status(400).json({ error: err });
+        console.log(err);
+    }
+});
+
+//sign in route will issue a new sessionToken when the old one has expired
+authRouter.post('/signin', async (req, res) => {
+    try {
+        
+        const user = await User.findOne({
+            email: req.body.email.toLowerCase(),
+        });
+
+        // Check if user exists BEFORE accessing password
+        if (!user) {
+            return res.status(400).json({ error: 'No user found' });
+        }
+
+        //if user exist, check pw
+        if (!bcrypt.compareSync(req.body.password, user.password)) {
+            return res.status(400).json( { error: "incorrect password"});
+        }
+            const sessionToken = jwt.sign(
+            {_id: user._id},
+            process.env.JWT_SECRET,
+            //expires in 1 day
+            { expiresIn: 60 *60 *24 }
+        );
+        res.status(200).json({ message: "Login Successful", sessionToken, user: user });
+
+    } catch(err) {
+        res.status(400).json({error: err});
+        console.log(err);
+    }
+});
+
+export default authRouter
