+++ server/server.js
@@ -3,7 +3,9 @@ import express from 'express';
 import cors from 'cors';
 import 'dotenv/config';
 import mongoose from 'mongoose';
-import userRouter from './routers/userRouter.js';
+import authRouter from './routers/authRouter.js';
+import staffRouter from './routers/staffRouter.js'
+import userRouter from './routers/userRouter.js'
 
 const app = express();
 const PORT = process.env.PORT || 5000;
@@ -11,6 +13,8 @@ const PORT = process.env.PORT || 5000;
 // Middleware
 app.use(cors());
 app.use(express.json());
+app.use(authRouter);
+app.use(staffRouter);
 app.use(userRouter);
 
 // Routes
